requirements.txt
Django>=4.0,<5.0
mysqlclient>=2.0.0
djangorestframework>=3.12.0
djangorestframework-simplejwt>=5.0.0
Pillow>=9.0.0
django-cors-headers>=3.10.0
